[Chore] LoginView와 LoginViewController 분리
